better errors
